+++ pulsar-sql/presto-pulsar-plugin/src/main/resources/META-INF/services/com.facebook.presto.spi.Plugin
@@ -1,1 +1,1 @@
-com.facebook.presto.pulsar.PulsarPlugin
+org.apache.pulsar.sql.presto.PulsarPlugin
