--- pulsar-sql/presto-pulsar/src/main/java/com/facebook/presto/pulsar/PulsarConnectorFactory.java
@@ -1,72 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package com.facebook.presto.pulsar;
-
-import com.facebook.presto.spi.ConnectorHandleResolver;
-import com.facebook.presto.spi.connector.Connector;
-import com.facebook.presto.spi.connector.ConnectorContext;
-import com.facebook.presto.spi.connector.ConnectorFactory;
-import com.google.inject.Injector;
-import io.airlift.bootstrap.Bootstrap;
-import io.airlift.json.JsonModule;
-import io.airlift.log.Logger;
-
-import java.util.Map;
-
-import static com.google.common.base.Throwables.throwIfUnchecked;
-import static java.util.Objects.requireNonNull;
-
-public class PulsarConnectorFactory implements ConnectorFactory {
-
-    private static final Logger log = Logger.get(PulsarConnectorFactory.class);
-
-    @Override
-    public String getName() {
-        return "pulsar";
-    }
-
-    @Override
-    public ConnectorHandleResolver getHandleResolver() {
-        return new PulsarHandleResolver();
-    }
-
-    @Override
-    public Connector create(String connectorId, Map<String, String> config, ConnectorContext context) {
-        requireNonNull(config, "requiredConfig is null");
-        log.debug("Creating Pulsar connector with configs: %s", config);
-        try {
-            // A plugin is not required to use Guice; it is just very convenient
-            Bootstrap app = new Bootstrap(
-                    new JsonModule(),
-                    new PulsarConnectorModule(connectorId, context.getTypeManager())
-            );
-
-            Injector injector = app
-                    .strictConfig()
-                    .doNotInitializeLogging()
-                    .setRequiredConfigurationProperties(config)
-                    .initialize();
-
-            return injector.getInstance(PulsarConnector.class);
-        } catch (Exception e) {
-            throwIfUnchecked(e);
-            throw new RuntimeException(e);
-        }
-    }
-}
