--- pulsar-sql/presto-pulsar/src/main/java/com/facebook/presto/pulsar/PulsarConnectorUtils.java
@@ -1,40 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package com.facebook.presto.pulsar;
-
-import org.apache.avro.Schema;
-import org.apache.pulsar.client.admin.PulsarAdmin;
-import org.apache.pulsar.client.admin.PulsarAdminException;
-import org.apache.pulsar.common.naming.TopicName;
-
-public class PulsarConnectorUtils {
-
-    public static Schema parseSchema(String schemaJson) {
-        Schema.Parser parser = new Schema.Parser();
-        return parser.parse(schemaJson);
-    }
-
-    public static boolean isPartitionedTopic(TopicName topicName, PulsarAdmin pulsarAdmin) {
-        try {
-            return pulsarAdmin.topics().getPartitionedTopicMetadata(topicName.toString()).partitions > 0;
-        } catch (PulsarAdminException e) {
-            throw new RuntimeException(e);
-        }
-    }
-}
