--- pulsar-sql/presto-pulsar/src/main/java/com/facebook/presto/pulsar/PulsarTableLayoutHandle.java
@@ -1,44 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package com.facebook.presto.pulsar;
-
-import com.facebook.presto.spi.ConnectorTableLayoutHandle;
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import static java.util.Objects.requireNonNull;
-
-public class PulsarTableLayoutHandle implements ConnectorTableLayoutHandle {
-    private final PulsarTableHandle table;
-
-    @JsonCreator
-    public PulsarTableLayoutHandle(@JsonProperty("table") PulsarTableHandle table) {
-        this.table = requireNonNull(table, "table is null");
-    }
-
-    @JsonProperty
-    public PulsarTableHandle getTable() {
-        return table;
-    }
-
-    @Override
-    public String toString() {
-        return table.toString();
-    }
-}
