--- pulsar-sql/presto-pulsar/src/main/java/com/facebook/presto/pulsar/PulsarTopicDescription.java
@@ -1,68 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package com.facebook.presto.pulsar;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import static com.google.common.base.MoreObjects.toStringHelper;
-import static com.google.common.base.Preconditions.checkArgument;
-import static com.google.common.base.Strings.isNullOrEmpty;
-import static java.util.Objects.requireNonNull;
-
-public class PulsarTopicDescription {
-    private final String tableName;
-    private final String topicName;
-    private final String schemaName;
-
-    @JsonCreator
-    public PulsarTopicDescription(
-            @JsonProperty("tableName") String tableName,
-            @JsonProperty("schemaName") String schemaName,
-            @JsonProperty("topicName") String topicName) {
-        checkArgument(!isNullOrEmpty(tableName), "tableName is null or is empty");
-        this.tableName = tableName;
-        this.topicName = requireNonNull(topicName, "topicName is null");
-        this.schemaName = schemaName;
-    }
-
-    @JsonProperty
-    public String getTableName() {
-        return tableName;
-    }
-
-    @JsonProperty
-    public String getTopicName() {
-        return topicName;
-    }
-
-    @JsonProperty
-    public String getSchemaName() {
-        return schemaName;
-    }
-
-    @Override
-    public String toString() {
-        return toStringHelper(this)
-                .add("tableName", tableName)
-                .add("topicName", topicName)
-                .add("schemaName", schemaName)
-                .toString();
-    }
-}
